Move package out of detectors and fix lint errors
