+++ ner_v2/detectors/temporal/date/en/data/properties.csv
@@ -1,3 +1,3 @@
-property,prefixes,suffixes,bot_message_contexts,prefix_within_n_words,suffix_within_n_words,prefix_suffix_both_required
-arrival,arriving|arrive|return|returning at|will be back on|traveling back|coming back|returning back|returning on|return date|arrival date|check out date|check-out date|date of check-out|check out|return date:|return date -,,arriving|arrive|return|returning at|will be back on|traveling back|coming back|returning back|returning on|return date|arrival date|check out date|check-out date|date of check-out,4,0,false
-departure,onward date\:|onward date -|departure date|leaving on|starting from|departing on|departing|going on|for|departs on|traveling on|going on|starting on|departure date|date of travel|check in date|check-in date|date of check-in|date of departure\.|,,traveling on|going on|starting on|departure date|date of travel|check in date|check-in date|date of check-in|date of departure\.|,4,0,false
+property,prefixes,suffixes,bot_message_contexts,prefix_suffix_both_required
+arrival,arriving|arrive|return|returning at|will be back on|traveling back|coming back|returning back|returning on|return date|arrival date|check out date|check-out date|date of check-out|check out|return date:|return date -,,arriving|arrive|return|returning at|will be back on|traveling back|coming back|returning back|returning on|return date|arrival date|check out date|check-out date|date of check-out,false
+departure,onward date\:|onward date -|departure date|leaving on|starting from|departing on|departing|going on|for|departs on|traveling on|going on|starting on|departure date|date of travel|check in date|check-in date|date of check-in|date of departure\.|,,traveling on|going on|starting on|departure date|date of travel|check in date|check-in date|date of check-in|date of departure\.|,false
