+++ ner_v2/detectors/temporal/date/en/data/range_properties.csv
@@ -1,3 +1,3 @@
-property,prefixes,suffixes,bot_message_contexts,prefix_within_n_words,suffix_within_n_words,prefix_suffix_both_required
-start,from,,start|starting from,1,0,false
-end,to|till,,end|ending,1,0,false
+property,prefixes,suffixes,bot_message_contexts,prefix_suffix_both_required
+start,from,,start|starting from,false
+end,to|till,,end|ending,false
