+++ ner_v2/detectors/temporal/date/hi/data/properties.csv
@@ -1,1 +1,1 @@
-property,prefixes,suffixes,bot_message_contexts,prefix_within_n_words,suffix_within_n_words,prefix_suffix_both_required
+property,prefixes,suffixes,bot_message_contexts,prefix_suffix_both_required
