+++ ner_v2/property_assignor/property_assignor.py
@@ -3,9 +3,9 @@ from __future__ import absolute_import
 import regex as re  # We are using regex because it provides accurate word boundaries than built in re module
 
 import ner_v2.detectors.utils as v2_utils
-from ner_v2.detectors.property_assignor.fields import AssignedProperty
-from ner_v2.detectors.property_assignor.reader import PropertiesDataReader
-from ner_v2.detectors.property_assignor.utils import clean_text
+from ner_v2.property_assignor.fields import AssignedProperty
+from ner_v2.property_assignor.reader import PropertiesDataReader
+from ner_v2.property_assignor.utils import clean_text
 
 
 class PropertyAssignor(object):
